+++ sat/cli/bootsys/util.py
@@ -22,6 +22,7 @@ ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
 OTHER DEALINGS IN THE SOFTWARE.
 """
 
+from collections import defaultdict
 import logging
 import re
 import shlex
@@ -32,9 +33,24 @@ LOGGER = logging.getLogger(__name__)
 
 
 class RunningService:
-    def __init__(self, service, dry_run=True):
+    def __init__(self, service, dry_run=True, sleep_after_start=0):
+        """
+        Create a new context manager that starts a service in a context and
+        stops it when leaving the context.
+
+        Note that if a service was already running, it will be stopped after
+        the context is exited. It does not leave the service in the state it
+        was in before entering the context.
+
+        Args:
+            service (str): name of the service to start/stop
+            dry_run (bool): if True, do not really start and stop the service.
+            sleep_after_start (int): number of seconds to sleep after starting
+                the service when entering the context.
+        """
         self._service = service
         self._dry_run = dry_run
+        self._sleep_after_start = sleep_after_start
 
     def _systemctl_start_stop(self, state):
         """Start or stop the given service.
@@ -53,6 +69,11 @@ class RunningService:
 
     def __enter__(self):
         self._systemctl_start_stop(True)
+
+        if self._sleep_after_start:
+            LOGGER.info("Sleeping for %s seconds after starting service %s.",
+                        self._sleep_after_start, self._service)
+            time.sleep(self._sleep_after_start)
 
     def __exit__(self, exc_type, exc_value, exc_traceback):
         self._systemctl_start_stop(False)
@@ -98,52 +119,30 @@ def get_ncns(groups, exclude=None):
     """
     all_groups = get_groups(groups)
     excluded = get_groups(exclude or [])
-    return sorted(all_groups - excluded)
+    return all_groups - excluded
 
 
-def wait_for_nodes_powerstate(nodes, state, timeout):
-    """Wait for nodes to reach the given power state.
+def k8s_pods_to_status_dict(v1_pod_list):
+    """Helper function to convert a V1PodList to a dict.
 
-    If not all nodes in the given state after `timeout` seconds, then
-    a set is returned containing the nodes that did not reach the
-    given state.
+    It is used for recording the pod phase for every pod in the system.
+
+    The dict has the following schema:
+    {
+        namespace (str): {
+            name_in_namespace (str): pod_phase (str)
+        }
+    }
 
     Args:
-        nodes ([(hostname, Command)]): a list of 2-tuples with
-            the hostname in the first position and a corresponding
-            pyghmi.ipmi.Command object.
-        state (str): a power state (either 'on' or 'off')
-        timeout (int): a timeout, in seconds, for the given nodes
-            to reach the desired state.
+        v1_pod_list: a V1PodList object from the kubernetes library
 
-    Return:
-        a set of nodes that haven't reached the given state (if
-        successful, this is the empty set.)
+    Returns:
+        a defaultdict in the above form
 
     """
-    start_time = time.monotonic()
-    completed_nodes = set()
+    pods_dict = defaultdict(dict)
+    for pod in v1_pod_list.items:
+        pods_dict[pod.metadata.namespace][pod.metadata.name] = pod.status.phase
 
-    while True:
-        remaining_nodes = set()
-        for node, cmd in nodes:
-            if node in completed_nodes:
-                continue
-
-            if cmd.get_power()['powerstate'] == state:
-                completed_nodes.add(node)
-            else:
-                remaining_nodes.add(node)
-
-        LOGGER.debug('Remaining polled nodes: %s', remaining_nodes)
-
-        if not remaining_nodes:
-            return remaining_nodes
-
-        elif time.monotonic() - start_time > timeout:
-            LOGGER.error("Reaching power state '%s' on nodes timed out after %d seconds: %s",
-                         state, timeout, ", ".join(remaining_nodes))
-            return remaining_nodes
-
-        else:
-            time.sleep(1)
+    return pods_dict